remove success comment and add stage label
It looks like actions that do a merge will not trigger the `push` event.
So unfortunately it doesn't look like we can rely on the existing
functionality to add the stage label.

Remove the previously added successful merge comment and just add the
staging label manually to denote to the developer that it was merged
successfully and should be on staging soon.