fix merge to staging workflow issues
Fix issues with the current implementation where if there was a merge
conflict or if a branch was already merged it would leave the
`deploy-staging` label and would mark the action as failed without a
great way to retry in order to fix the error.

Now if there is a merge conflict or a branch is already merged it will
log a comment and remove the `deploy-staging` label regardless of the
action result.